feat: Mypage API/2024-12-31

+++ src/main/java/com/ohgiraffers/ukki/user/model/dao/MypageMapper.java
@@ -5,5 +5,5 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MypageMapper {
-    MypageDTO findUserInfoByUserNo(Long userNo);
+    MypageDTO findUserInfoByUserNo(int userNo);
 }
